+++ payroll-backend/src/main/java/org/eshames/payroll/payrollbackend/model/taxes/Tax.java
@@ -3,5 +3,5 @@ package org.eshames.payroll.payrollbackend.model.taxes;
 import org.eshames.payroll.payrollbackend.model.dto.impl.TaxDTO;
 
 public interface Tax {
-    float calculateTax(float taxableIncome, TaxDTO taxDTO);
+    double calculateTax(int taxableIncome, TaxDTO taxDTO);
 }
